fix: life worker reports wrong value
There is a race when the worker is shutting down between the local
life value and what the actual life of the entity is at. If something
called Check() during that period then it would report the wrong value.

The solution is to always update the entity life value even if we're
about to exit the loop. This ensures that anything checking the current
value outside of the loop in another goroutine will get the correct
value.