feat: enable finding oci subnets at model creation
Previously subnets etc were created when the first instance in a model
was created. With DQLite, Subnets is called, via ReloadSpaces, at model
creation to populate subnets and availability zones. Now that
referential integrity exists, the AZ must exist for a machine's hardware
characteristics to be saved at deployment.

Solve the problem by overloading the environs method: Subnets, to create
the subnets, availability zones, et al if no subnets are found when
called.