feat: ensure death
This ensures that all workers are correctly killed if they encounter
a error. This is a bit heavy handed, but it does ensure that all
model worker managers are correct terminated. Once a model becomes dead
that's it, stop using it.

It's possible for a worker to hang on until it's done with it's work,
but it must start doing that work at the dying phase.