feat: handle dying dbs as soon as possible
If the database is being removed, correctly handle that the database
is dying. This ensures that at the first operatunity we handle the
database is dying. It then prevents not found or similar error messages
about entities not existing. It won't remove them completely but it
does remove a lot of them.

The solution is to know when a tracked db is dying and just return
the DBNotFound error message. If the service or client want's to retry
they can do so.