feat: start the removal of dummy provider
The dummy provider was just a testing provider with a really bad name.
We don't want to create the provider just by importing the path. This
creates an implicit dependency and is very magical. Instead, we want
to slowly remove this.

All tests should either use the real thing, or mocks.