feat: allow two signal channels to be merged
When accessing a worker from the cache, it's useful to know when to
give up. Considering both the context.Done and catacomb.Dying are
signal channels (nothing is sent, just that they can be closed to
signal completion) then we can merge them and watch for when one is
closed. First one to close the channel will signal the other death.