feat: drain agent objectstore
This was originally left out of the objectstore drainer as it was
becoming rather large. Instead, draining the agents objectstore is
actually rather trivial. We don't allow this to run in parallel of
other models, just because it's hard to manage. Yet, we can do it
upfront and then drain that, before moving on.