feat: initial offer domain
Set out the domain structure including a stub for GetOfferUUID and a
basic implementation of Offer service methods. Offer contains code
pushed down from the facade and may not be complete. It'll be finalized
when the state portions are implemented.

State is setup with both model and controller database access as offer
access is persisted there. Stubs for calls required by the Offer service
method are setup.