feat: implement Offer and ModifyOfferAccess facade methods
These methods have been reimplemented using domain services.

This facade is a multi model facade. Access and model functionality are
via the controller database. Offers are contained in individutal model
databases.

Checking permission in both method methods is atypical. The model for
the facade is not always required by the call due to the original setup
of it.

For Offer, the minimal requirement that the api caller has superuser
access on the controller or model admin access for the model being
acted on. Same with the owner specified if different from the api
caller. This is to faciliate JAAS use of the API.

For ModifyOfferAccess, the api caller meet the same requirements as
Offer, or having admin access on the offer.