feat: spread
Start introducing spread for juju testing. This creates two backends;

 - lxd is the local development backend
 - github ci will be for running the ci jobs inside

The infrastructure for the backends and ensuring that we've got juju
installed is most of the work.

There is an issue where we're not waiting for a charm to be installed
before tearing down. That can be done in a follow up PR.