feat: status history for controller model
We are unable to insert the status history for a controller model
at the time of creation, so instead we need to do it in the bootstrap
worker. This aligns the controller model with all other models. In
fact the time used for insertion is actually correct, because the
status indicates when the controller model is "available" which is
indeed correct when we write the status.