feat: namespace global worker loggers
Some workers operate globally - they work in isolation in the service
of others to fulfill the end goal. These workers provide access to
workers for namespaces (literal "controller" name or a model UUID), that
provide sub units of worker and do this concretely via access to the
underlying runner worker. The problem is that it is sometimes hard to
know which worker is reporting what using debug-log or the
/var/log/juju/machine-N.log. To help with this, every
new logger will now be suffixed with the model uuid or controller
name. There is a downside to this, if we have lots of models without
restarting could lead to unbounded growth. For now this should suffice,
but there is very much a flaw at the heart of the loggo package.