feat: watcher summaries
When inspecting or debugging watchers within the changestream it is
handy to know what subscription is for what purpose. This does that
by ensuring that all watchers are forced to tell us their intent.
Then we can see from a highlevel what's happening. This could help
identify slow watchers in production.