feat: remove resources in facades
We no longer require resources for facades, instead we can use the
more robust watcher registry. The watcher registry uses the worker
runner to manage watchers served from the facades. It provides
a cleaner abstraction for managing workers, and guarantees that only
one worker for a given name is running at a given time.

This change also removes the issue of which watcher registry to use,
either the new watcher registry or the resource backed one. It's fairly
evident that the resources concept didn't take off. It had some nice
properties, but was only half baked.