feat: implement ListApplicationOffers with services
The new implementation of getApplicationOffersDetails does not try to
accomplish all prior functionality to begin with. The logic around
permissions was convoluted. It will be made more clear as more facade
calls are implemented with services.

The ListApplicationOffers implementation is not intuitive, but does
match what is done in 3.6. It does require the user auth tag to have
superuser permission on the controller or admin permission on the model.
Only the current user is listed for the found offer, even for an admin.