refactor(application): handle arch correctly
When deploying an application, by the time we get to the state layer we
need know what the architecture is. Unknown, or a missing value, lead to
foreign key violations when filling in the application_platform
(correctly)

However, we are a little loose with how we handle the architecture. This
is a result of some complications:
- The arch is present in two places in params, the constraints and the
  platform, so we don't really have a single source of truth.
- We create applications with the "Deploy" and "DeployFromRepository"
  facades, which drive the service in slightly different ways.
- We also need to default the arch to model constraints, and then a
  hard-coded value.
- We need to know this resolved arch to be able to get the correct charm
  from charmhub, in the case of DeployFromRepository

I have added a new service method which resolves constraints, including
the arch. This is then called by DeployFromRepository to get this arch.
This isn't ideal. However, soon we are re-writing the api, at which
point we will push a lot of DeployFromRepository into service methods.