feat: return the same registry
No matter how many times you call get watcher registry, it will always
return the same one. This is a bit clunky as the API server should
never call GetWatcherRegistry with the same connection id, for the
lifetime of the process. Restarting the process and starting at an
offset is fine.

Tests are added to ensure no races exist.