feat: provide a unique namespace
When a process restarts, it's possible that the client doesn't witness
this. In this case you'll have watcher ids that don't match the
existing watchers. This will lead to very strange errors indeed.
Instead, we can just namespace them to a unique value on start up and
prefix every watcher request with that prefix. Thus, you'll get the
correct not found error upon restart.