feat: prevent go routine leak
Passing a worker into the worker runner can leak if the runner is
killed before it's fully onboarded. This is because the start worker
method should create the worker there and then.

The solution is to wait for the runner to take ownership during the
registration period, before return back. This might be slower, but we
can guarantee that nothing leaked.