feat: remove testing code from prod code
The remote relations worker tests used a config runner to allow the
getting of a application worker. This was just all ceremony and it
was caused by the fact that it was doing the wrong type of testing.
If you want to look into the type at test time, just do it!

Also fixed a lot of static analysis errors before making more agressive
changes.