fix(subordinate): deploy bug
As we know, charm endpoints have a 'scope' property, which is set to
either 'global' or 'container'.

However, it turns out that we also implicitly treat relations themselves
as if they have a scope property, which isn't just equal to the scope of
either of the charm endpoints it relates.

You can see this in Mongo state represented here:
https://github.com/juju/juju/blob/610241d8cd49ddc9ee76a5e7b7c0c382383576df/state/state.go#L2142-L2145

However, since we now have referential integrity and look up the charm
endpoint scope (and previously, the relation endpoint scope, which is
not the same thing) by joining on our charm metadata, we cannot account
for this behaviour with the current schema.

This lead to a bug deploying subrodinates, where the uniter can't find
any alive container scoped relations for the subrodinate, and assumes it
should be terminated.

The solution was to impart an intrinsic scope to relations themselves,
calculated from the scopes of the charm endpoints they relate to, and to
store this on the relation itself in the db. So now, when it comes to
representing a relation endpoint, we use the scope of the relation
itself, not the charm endpoint.