feat: allow sqlite to build
This builds sqlite to known good version. We can't for some reason
use anything beyond this point. I will need to investigate why, the
error codes from the compilation doesn't state any reason why!

For now, bump this up a tiny bit - this is ahead of the noble version
in production, but development in this case should be ensuring that
we're not caught short.