refactor: remove the concept of storage provider type
Within the storage and application domains we held on to the fact that
storage could be provisioned either directly against a storage provider
type or that of a pool but you can not be provisioned against both.

This modeling of provisioned storage from one of two sources was in
reality an artifact of how the Juju API was constructed over time. It
lead to the need of having to constantly check if one of two values was
set and then have two different code paths based on the value.

Instead of this approach we now consider all storage to only be
provisioned against a storage pool. A pool is then always associated
with a provider.

This allows Juju to have strong RI in the data model and a clear
explanation that storage is always from a single pool.

A follow up commit will be done to establish a storage pool for each
provider in the model "a default pool".