feat: get the api info for model
Getting the APIInfo for the model needs to happen over the getting
of the domain services. This is late bound and cached via the runners.
Once an api connection is broken it will attempt to request the
info again. Thus we only ever have one connection to a controller
for **all** models.

Tests have been added to ensure this is the case.