feat: handle connections and redirects
When getting a connection, we need to handle redirects. This does this
when making a connection. If the connection fails when attempting to
update the external controller cache, we just ignore the error and
return the connection. We don't care if we have to perform the
update to the external controller again. The next time we connect will
perform that for us.