feat: implement api info getter
The APIInfoGetter is responsible for getting the model information
from a local model, otherwise look up a redirection or external
controller model.

This is a replication of apiserver/common ControllerAPIInfoForModels
method. We don't need to call this over the API if we're performing
the request locally. This saves on a lot of API requests.