feat: separate worker construction
When creating a remote application worker, we don't actually care what
worker we have, just that it takes a config and returns a worker. This
breaks the coupling between the workers. Thus we can better unit test
each worker, as they don't depend every worker at every level. There
is still a need for a test against all workers, but not for every
permutation.