feat: remove offeruuids from worker
The worker runner already has the workers with the offer uuids, so
there is no point having another map, attempting to keep in sync with
the workers. This allows us to simplify and focus on what the purpose
of the worker, which is to spawn and track application workers. Nothing
more.