feat: new storage providers for Kubernetes
Juju has always supported tmpfs and rootfs storage providers for
Kubernetes but the way this has been done is by composing the machine
providers into Kubernetes and trying to make them understand both views
of the world.

It is also up to the buisness logic of Juju to understand that these
extra storage providers are available for the model's environ. It is
much easier to have the environ tell the logic what can be used.

We now want to make each environ responsible for controlling their
storage providers. This change will help remove more case's of if then
else checks for Kubernetes.

While here I also removed legacy errors from the storage files in the
Kubernetes provider.