refactor: remove usage of common storage pools
This commit removes the usage of common storage pools from state testing
in the storage domain. We doing away with this common understand of
builtin storage pools in the model and pushing this back on to the
provider.

Additionally to this the state tests should have been making dummy pools
in the model instead of performing integration across package
boundaries.