fix: subordinate unit status
Prior to this change, we did not show subordinate units in juju status

This was because we forgot to fill in all the attributes of a ret vars
struct, meaning data we later needed was not present.

Fix this by filling in this data.

As a flyby, make some minor optimisations to the facade status
processing code.