feat: actions DDL
An operation is when a charm action is run, or a command is run on a
machine or unit. Action operations are linked to a charm action.
Operations have a status.

A task represents individual receivers of the operation being run. It
can be a unit or a machine. It has status and a log of messages.

The output of each task is stored in the object store.