feat: split consumer and offerrer workers
Split the workers so we're only dealing with the context at hand.
The prior worker dealt with both sides of the connection which made
it VERY confusing to understand what was going on. We're no longer
doing this. With the introduction of the domain services we want to
stop doing that. We can then write exact code for both of them. If
there are common sub workers between them, then we'll deal with that
at the time.