refactor: rewmove storage registry from bootstrap
This task removes the storage registry from the bootstrap worker. The
reason for doing this was the it is the domain which will be responsible
for inserting any default storage pools into the model.

Doing it this way allows the bootstrap worker to just call domain
services in the model and not have to rely on another dependency.

A follow up commit is comming to introduce the new domain func for
adding default storage pools.