refactor: remove common storage pools from tracked worker
The storage registry tracked worker would automatically install the
controller's builtin storage provider with that of the environ. We are
moving away from this so the environ now offers up all providers that it
supports.

This allows us to remove a lot of if then else logic around Kubernetes.

This change allows the environ to be responsible end to end for the
storage providers in the model.