refactor: explicit registration of machine storage provisioners
This changes how the storage provisioner works on machines. Before this
change the storage provisioner assumed that the common storage provider
types were all available to run on machine storage provisioners.

This relationship was assumed and not offered by any formal contract.
Instead of this the provisioning worker is now explicitly registered
with the providers that it supports.

This allows us to continue to also remove usages of common storage
providers.