test: remove usage of common storage providers in migration tests
This change removes the usage of common storage providers from the model
migration target tests. The reason for doing this is this is an
integration test that can't be performed.

If the test starts to care about available providers in the model then
it should be seeded with testing providers.

Using providers that can alter the system and not be inspected in a test
is fraught with danger.