refactor: remove common storage providers and references
This commit removes all usages of the common storage providers that were
left. This was mostly tests which have been replaced with nil values.
All of these tests were not actually using the values but simply
referring to them to make the test compile.

As part of this work the internal storage provider package now exports
the builtin providers so that they can be referenced and used from the
environs.

This allows for the environ to start composing the set of providers that
it supports.

While here I also updated loopfs, tmpfs & rootfs so that they now offer
default storage pools for their provider to the model.