refactor: let environs decide the storage providers and pools available
This commit introduces a new change to the Juju model. Previous to this
the storage providers offered by an environ were augmented further in
the modeling of Juju with the concept of builtin providers.

Over time we have ended up with the case where not all builtin providers
are supported for every environ and so a dance has to be performed to
make sure the set of provider types is correct for the environ and
model.

By having the environ tell the caller what it supports this need goes
away entirely and the set is always correct.

On top of this each environ storage provider is also making a default
pool available for their default storage provider that matches the same
name as the provider.

This will allow for users to always be creating storage against a pool
and not a provider type. This will remove the double meaning of values
in our model and create an explicit model where all storage is against a
pool.