fix: discrepancy between charm reference name and name
For the vast majority of charms, the name of the charm in charmhub is
the same as the name of the charm in it's metadata. However, there are
exceptions (and it's a supported use case).

However, sometimes we conflate the two. For example, when constructing
the charm url of a 'domainCharm' in the apiserver, we would use the
charm metadata name, but really this should (arguably) be the reference
name. However, this is not clear.

Anyway, I have fixed this by removing the URL method from domainCharm,
so the question is moot. Instead, we use the charm identifier returned
from the domain service, instead of attempting to reconstruct it.