feat: allow get db to fail
It's entirely possible to block the dbaccessor loop, based on one
bad actor. This can happen if the underlying database driver doesn't
fulfill the request to open a database. There isn't anything we can
do in this instance, other than to retry again if it fails. The problem
might be alleviated by the incoming delete database pragma, but there
is still a possibility of this happening even if that is resolved.