feat: request cancellation stop open
When the request was cancelled it didn't stop the opening of the app
it just stopped waiting for a response from opening the db.

This change passes the context to the application and ensures that
we don't block other requests to open.

More importantly, we're not blocking the main thread. We're doing as
little work as possible, before unblocking the main thread.