feat: changestream pruner
The changestream pruner does not handle the removal of models very well.
This is because the strategy has evolved since this was written. It
is expected that each model should be a worker runner and handle
the database going missing. This is inline with other workers (i.e.
the removal worker).

This should ensure that the we don't prevent the leaking of the change
log items when a model becomes dead.