refactor: only return offer users when admin access required
Updated getApplicationOffersDetails to require a permission level in
anticipation of implementing FindApplicationOffers which uses the same
funcationality as ListApplicationOffers.

This change is also a departure from juju 3 behavior as
FindApplicationOffers now required at least read access to the model.
Allowing the method to be called without any permission appeared to be a
failure in the prior implementation.