test: fix bootstrap storage command test
There existed a bootstrap command test that was asserting a new pool
could be created for the loop provider. It assumed the loop provider was
always available.

In tests we have the dummy provider that dictates the storage providers
available. It also has a set of inbuilt provider type which should be
being used.

This fixes the test by now using the provider types offered by the dummy
provider.