test: fixes tracked worker tests
The tracked worker used to add the builtin storage providers to each
environ registry when the caller requested the provider types. This
essentially was the injection mechanism used to get these provider types
into the model.

Now that the environ's report this themselves we have had to update the
tracked worker tests so that are no longer asserting this injection.