refactor: remove storage pool origin builtin
We introduced the concept of a builtin storage pool to represent pool's
that were built into the Juju controller. Since then we have changed
this model around a bit by where this concept has been removed and the
environ is now in charge of dictating the full set of providers that are
on offer in the controller.

Because of this all storage pools now come from environ's and the
concept of builtin doesn't anymore.