refactor: change storage migration to not care about builtins
Storage migration was not exporting builtin storage pools. As we have
removed the concept of builtin pools and we are not currently supporting
model migration I have made this service use the list all method.

This is how we want migration to work going forward for storage but a
final design has not been finalised.