refactor: use string when passing storage pool uuid to state
In the model when seeding default storage pools into a new model we were
passing the strong type for a storage pool uuid down to state. Our
standard now is pass via primitive type.

I am leaving the origin as a strong type as this is related to the DB
and is a primitive type.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>