feat: default storage pool uuid map
This commit introduces the concept of default storage pool uuid for
storage pools that are created by the model's environ.

Until now we have been generating a new storage pool uuid for every
default pool in the model. This means that between models and migration
storage pool uuids will change.

We would like a system where storage pools that are fixed stay consitent
across controller's and model's. To do this we have introduced a static
map that translates the default storage pools on offer by each environ
with a fixed storage pool uuid.

The method to achieve this is not ideal for every case but we believe
that this is a good first step moving forward. It allows us to build on
the concept over time.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>