refactor: use constant default storage pool uuids in model creation
When creating a new model we seed a set of default storage pools into
the model. Until now we had been generating a new storage pool uuid for
each of the default pools.

This change now makes the model service use the set of statis uuids
available ensuring that default pools are always created with the same
uuid. If a new default pool comes into the set but does not have a
mapping for it's uuid we log a warning and disregard the operation.

This is the safest approach and this problem "should" be caught in
testing before a release of Juju.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>