Ignore NotFound error if not using port security
The bootstrap workflow ends up querying Neutron for the controller
security group regardless of whether port security was enabled. If port
security is disabled, the security group is missing and the call to
`getSecurityGroupByName` will fail.

The change introduces a check of the error returned by
`getSecurityGroupByName` and turns it into a `NotFound` error so that
the caller, `OpenModelPorts` can ignore it.

Closes: https://github.com/juju/juju/issues/20513
Signed-off-by: Nicolas Bock <nicolas.bock@canonical.com>