feat: wire the action facades with the new operation domain
This patch wires the action facades (Actions() and Cancel()) with the
newly created (and stubbed) operation domain.
The main functional difference is that both the list and cancel of
actions is done in a bulk fashion, thus avoiding a N+1 query
problem.

As a flyby, the core/actions packaged was renamed as action for
consistency and a UUID struct + methods for action have been added.