fix(remove): fix removal of unit whilst machine pending
fixes #20467

There was a bug in the deployer/uniter/removal domain which meant that a
unit whose machine was still pending would deadlock if it was removed.

This is because of two compounding factors:
- The removal domain waits until the uniter has marked a unit as dead
  before removing it
- The deployer doesn't start the uniter for units that are dying

This means, if the unit is removed beforew it's uniter starts, it will
never be marked as dead.

The way we resolve this is by fixing the deployer to be able to mark a
unit as dead as well. This was actually how 3.6 behaved. See:

https://github.com/juju/juju/blob/2da8de8a4fd4f5dbbed42c6e5c355a37649f1d22/apiserver/facades/agent/deployer/deployer.go#L76

https://github.com/juju/juju/blob/2da8de8a4fd4f5dbbed42c6e5c355a37649f1d22/apiserver/common/remove.go#L56-L59

We can do this because we know that the deployer only calls unit.Remove
for units that are not alive, and guards against the uniter existing