feat: unix tuple index for volume attachment plans
This commit introduces a unique index on the
storage_volume_attachment_plan table so that only one attachment plan
can exist per volume and net node uuid.

This is done because the storage provisioning api as of today expects
that it can reference a single attachment plan by volume id and machine
uuid.

If the possability exists that two attachment plans exist for these two
values then it will produce unknown results within Juju.

When the api moves over to using uuid's to refer to entities in the
model we can look at removing unique index's like this.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>