chore: handle action not found error in api
This commit reverts the bulk service method signatures introduced in the
previous commit, as per requested in the PR review.

The service now gets called once per input arguments, this is
particularly suboptimal for the Cancel() endpoint which receives N input
arguments and results in N db hits.
This issue should be addressed when redesigning the API.

As a flyby, mock based tests were added to the action API for coverage.