feat: prevent worker lock ups
With the addition of context to the methods for accessing a worker
runner guarded resource, we need to ensure that we don't prevent the
loop from locking up on a write to a channel. The problem is that the
caller to dbaccessor.GetDB (as an example), cancels the call, the
worker loop is still running, attempting to send a error to a channel
that nobody is reading. This was totally fine when we didn't have a
context, as the tomb or the catacomb dying was the only thing that
could cancel the request. The solution is simple, we buffer the
channel. We know that we'll want to reiceve and error, if it's nil
or not, doesn't matter. The reality is that we should send the error,
thus unblocking the main loop.