refactor: only pass interface name when creating converting addresses
The instance-poller now just passes the interface name instead of the
whole interface when adding provider addresses to the network config
payload. This reflects the desire not to create address fields from the
interface, but to take them only from the listed addresses.

As a drive-by, a redundant if block is removed from the O7k provider.