refactor: application storage pool validator interface
The application domain service package had the concept of a storage pool
validator that allowed the application domain to ask questions of a
storage pool checking if a storage pool supported a given type of charm
storage.

While working on adding support for storage to IAAS models we need to
also ask further questions of the storage pools provider such as does it
use volumes or filesystems and what is the provisioning scope of the
provider.

The validator interface has most of what is required to support the
operations so it makes sense to widen the interface to be more of a
question source for the provider that backs storage pools.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>