feat: introduce storage kind to Juju DDL
This change introduces the concept of storage kind to the storage model.
We have not had this until now relying on the charm storage kind to do
this determination.

This falls down because there is no policy that says a storage instance
has to be the same kind as a charm. It is even possible for a instance
to not be related to a charm.

We also support bring storage in and out of the model that may not be
created the charm to start with.

For the moment I have left this value as non nil till all of the code
points can be touched up.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>