feat: compose new storage instances of their entities from charm storage
This commit brings in the ability for the application domain to now look
at the charm storage for which charm storage instances are being created
for. This allows for the correct composition of filesystem and volume's
under pinning a storage instance.

This work enable IAAS model's to start working again under Dqlite.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>