chore: remove the operation_exec table
This patch removes the operation_exec table, since the only contents are
actually operation parameters (command and timeout to be specific).
The outcome of this is that the operation type (action / exec) can be
deduced by checking if there's an entry in the operation_action table.

A wrong FK table was also fixed in the operation_task_status.