feat: model logger for provider tracker
Logs for model level activities should be in the model logs. This allows
the democratisation of the logs for owners of the model. Currently, all
activity with in a controller are sent to the controller logs and it is
only the model worker manager that can log to the model logs themselves.
This is a good step, but there are workers that work on behalf of the
model, that are not inside of the model worker manager. This introduces
the concept of allow that to happen. The intention is to not have
juju act as a black box for juju users that share a controller. Model
owners are more likely to look at the model logs than a controller
owner looking at the controller logs if they're managing multiple
controllers (from first hand experience).