feat: remove status tables
Remove the status tables as we can use the application and unit status
for each side. We'll have to ensure compatibility with the facades
when terminating, but it should be ok.

Additionally, we want the relation and not the relation endpoint. The
relation has all the information required.

Dropped any only "informational" fields, as they're impossible to
keep up to date after a migration.