fix: use `disk-type` instead of `type` when querying disk type on gce
When creating a storage pool on gce, the suggested command is:

   juju create-storage-pool mygpd gce type=pd-ssd

However, using `type` here causes a conflict: the pool manager
overwrites the `type` attribute with the storage provider type (`gce`
in this case). As a result, the storage pool is created, but the disk
type is set incorrectly.

This change replaces the `type` attribute with `disk-type`, allowing
users to correctly specify the desired disk type (e.g. `pd-ssd`)
upon instance creation.