feat: get ca cert
Move controller config CA cert out of the controller config and move it
into a better location. That means we can access the CA cert from the
controller domain. This has been a long standing issue where controller
config becomes a dumping ground for key value pairs. Here, we're going
to correctly locate the piece of information, but we're going to honor
the old API interface.

This should be done for all key, value pairs in controller config, but
we can do that over time.