feat: move jujuclient
The jujuclient aims to create a client for API for use by juju. This
is purely for API requests only, and should therefor be in the API
package.

The cmd package uses it mainly, but other things are using it. This
is because the ControllerStore is in the wrong location. Moving it
highlights this. For now this will suffice.