feat: reinstate auth offer context
The offer context for authorization was removed when we got rid of
state. We now need to put it back. This is the start of doing that.
It's very bare bones and doesn't have everything we need, but it
should work enough for getting the consume details for a macaroon.