refactor: handing of architecture
Resolves: #20559

There are a number of bugs in Juju that are caused by mishandling of
machine/charm architectures.

In particular, we allow Juju to deploy charms that only support arm to amd
(for example). Either regular `juju deploy`, but also `juju deploy --to`

This resulted from mishandling architectures. We include the charm we
want to deploy in the platform and constraints during deployment.

They are hardly ever both set, or both set to the same value. And to
make matters worse, different consumers read different values.

This is very bad modeling!

To add to the confusion, our charmhub resolver code does not correctly
set the architecture in the platform for charms in some cases. If we
specify a revision, charmhub will just return that revision (correctly)
irregardless of the architecture we ask for (or default to).

Resolve this by:
- Removing the requirement that we provide an architecture in platforms,
  and handle the case where we don't with sensible defaults. This allows
  us to keep track of whether we are constrained to amd64, or if we just
  want to default to it
- In cases where a charm revision supports only a single architecture,
  set the architecture constraint. This ensures that machines, and units
  for that application are always correct.
- When placing a unit/application, check in the service that the
  destination machine hardware characteristics do not conflict the
  constraints. So far, we only check the architecture, but more should
  be done in the future.