refactor: require machine placement to calc uuid and net node before call
When the machine placement state code was written it would calculate the
machine uuid and net node uuid in the state layer and then return these
new values to the caller.

From workflow perspective this has a design flow on effect in that it
makes other dependent things have to also perform their buisness logic
calculations down in state because the value is not known.

This isn't the design we utlimately want. We want to have these values
known ahead of time so that the vast majority of this type of work can
be done in the service layer.

The example being solved with this commit is storage. We need to node
net node uuids at the service layer.

This change now forces all callers of machine placement in state to
calculate these values up front and supply them in.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>