refactor: open up access to create machine args
There has been usages of create machine in the machine state outside of
this package and it relies on a struct type as input that was not
exported.

This commit exports the arg type that is requied by CreateMachine. These
usages shouldn't exist and has a deffinate smell about it. However this
line of work is not here to refactor the world so leaving as is for the
time being.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>