refactor: application service create uuids for state
This commit refactors the application domain service so that it now
generates the node node uuid for a unit. As well as this the service now
generates the uuid of the machine and also the net node uuid of any new
machines.

If a placement directive of a unit uses an already established machine
then these existing values will be substitued.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>