refactor: stub out CheckCharmPlacement
CheckCharmPlacement was used by refresh to perform a check that the
application is only deployed to a single architecture.

This isn't really a useful check. Really, charms should not support
multiple arches, but some do anyway. This check is a bit late in the
lifecycle.

Stub out this facade endpoint.