feat: create local bakery
This creates a new local bakery for creating and validating macaroons.
The idea is to greatly simplify the cross dependencies and improve
testing. It does mean that the bakery wrappers (LocalOfferBakery) are
doing more and the auth context is doing less. Yet it is pushing
things more to the edge, and we get a clear interpretation of what
should be happening.

One example in the old implementation is that we created and cloned
the underlying storage multiple times to ensure that the expiry time
was 3 minutes. We did this over and over again. At no point is anything
setting it to anything BUT 3 minutes!